refactor(backend): 拆分 API 接口定义到独立路由模块
- 将用户认证、模板管理和文件上传等接口分别拆分到独立的路由模块中
- 更新前端 API 调用路径以匹配新的路由结构
- 移除 main.py 中冗余的导入信息，简化代码结构
