+++ chat.md
@@ -39,3 +39,4 @@
 [songhahaha66][12:19 15:10:00]：文字没有正常显示，两个按钮应该在两侧
 [songhahaha66][12:19 15:15:00]：请使用多行文字列表，第二步界面少使用css
 [songhahaha66][12:19 15:20:00]：.step-actions {两个按钮在两头显示
+[songhahaha66][08-14 23:39:45]：项目规则文件：project_rule.md，请仔细阅读。你需要继续完成新的后端任务：现在如果所有api都放在main里边会显得太乱了，请你做一下拆分
