+++ frontend/src/api/auth.ts
@@ -27,7 +27,7 @@ export interface TokenResponse {
 
 class AuthAPI {
   private async request<T>(endpoint: string, options: RequestInit = {}): Promise<T> {
-    const url = `${API_BASE_URL}${endpoint}`
+    const url = `${API_BASE_URL}/auth${endpoint}`
     
     const config: RequestInit = {
       headers: {
